requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
boto3>=1.34,<2.0
supabase>=2.0,<3.0
gunicorn>=23.0,<24.0
